refactor: ratchet Tier 2 dependency boundaries
