+++ src/requirements-manager/checks/grafana-api.ts
@@ -6,8 +6,7 @@
  */
 
 import { config, hasPermission, getDataSourceSrv, getBackendSrv } from '@grafana/runtime';
-// eslint-disable-next-line no-restricted-imports -- [ratchet] ALLOWED_LATERAL_VIOLATIONS: requirements-manager -> context-engine
-import { ContextService } from '../../context-engine';
+import { fetchDataSources, fetchPlugins, fetchDashboardsByName } from '../../lib/grafana-api';
 import type { CheckResultError } from '../../types/requirements.types';
 
 /**
@@ -157,7 +156,7 @@ export async function hasDataSourceCheck(check: string): Promise<CheckResultErro
 export async function hasPluginCheck(check: string): Promise<CheckResultError> {
   try {
     const pluginId = check.replace('has-plugin:', '');
-    const plugins = await ContextService.fetchPlugins({ throwOnError: true });
+    const plugins = await fetchPlugins({ throwOnError: true });
     const pluginExists = plugins.some((plugin) => plugin.id === pluginId);
 
     return {
@@ -190,7 +189,7 @@ export async function hasPluginCheck(check: string): Promise<CheckResultError> {
 export async function hasDashboardNamedCheck(check: string): Promise<CheckResultError> {
   try {
     const dashboardName = check.replace('has-dashboard-named:', '');
-    const dashboards = await ContextService.fetchDashboardsByName(dashboardName, { throwOnError: true });
+    const dashboards = await fetchDashboardsByName(dashboardName, { throwOnError: true });
     const dashboardExists = dashboards.some(
       (dashboard) => dashboard.title.toLowerCase() === dashboardName.toLowerCase()
     );
@@ -307,7 +306,7 @@ export async function isEditorCheck(check: string): Promise<CheckResultError> {
  */
 export async function hasDatasourcesCheck(check: string): Promise<CheckResultError> {
   try {
-    const dataSources = await ContextService.fetchDataSources({ throwOnError: true });
+    const dataSources = await fetchDataSources({ throwOnError: true });
     return {
       requirement: check,
       pass: dataSources.length > 0,
@@ -331,7 +330,7 @@ export async function hasDatasourcesCheck(check: string): Promise<CheckResultErr
 export async function pluginEnabledCheck(check: string): Promise<CheckResultError> {
   try {
     const pluginId = check.replace('plugin-enabled:', '');
-    const plugins = await ContextService.fetchPlugins({ throwOnError: true });
+    const plugins = await fetchPlugins({ throwOnError: true });
 
     // Find the specific plugin
     const plugin = plugins.find((p) => p.id === pluginId);
@@ -414,7 +413,7 @@ export async function dashboardExistsCheck(check: string): Promise<CheckResultEr
 export async function datasourceConfiguredCheck(check: string): Promise<CheckResultError> {
   try {
     const dsRequirement = check.replace('datasource-configured:', '').toLowerCase();
-    const dataSources = await ContextService.fetchDataSources({ throwOnError: true });
+    const dataSources = await fetchDataSources({ throwOnError: true });
 
     if (dataSources.length === 0) {
       return {
