+++ src/requirements-manager/requirements-checker.utils.test.ts
@@ -6,7 +6,7 @@ import {
   validateInteractiveRequirements,
 } from './requirements-checker.utils';
 import { locationService, config, hasPermission, getDataSourceSrv, getBackendSrv } from '@grafana/runtime';
-import { ContextService } from '../context-engine';
+import * as grafanaApi from '../lib/grafana-api';
 import { getContentKey } from '../global-state/content-key';
 
 // Mock dom-utils functions with default return values
@@ -46,12 +46,10 @@ jest.mock('@grafana/runtime', () => ({
 }));
 
 // Mock ContextService
-jest.mock('../context-engine', () => ({
-  ContextService: {
-    fetchPlugins: jest.fn(),
-    fetchDashboardsByName: jest.fn(),
-    fetchDataSources: jest.fn(),
-  },
+jest.mock('../lib/grafana-api', () => ({
+  fetchPlugins: jest.fn(),
+  fetchDashboardsByName: jest.fn(),
+  fetchDataSources: jest.fn(),
 }));
 
 describe('requirements-checker.utils', () => {
@@ -217,7 +215,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should test specific data source configuration', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([
         { id: 1, name: 'Prometheus', type: 'prometheus', uid: 'prom-uid' },
         { id: 2, name: 'Loki', type: 'loki', uid: 'loki-uid' },
       ]);
@@ -236,7 +234,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when the health check reports a non-OK status', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([
         { id: 1, name: 'Prometheus', type: 'prometheus', uid: 'prom-uid' },
       ]);
 
@@ -254,7 +252,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when the health endpoint rejects (Grafana returns 400 on failed checks)', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([
         { id: 1, name: 'Prometheus', type: 'prometheus', uid: 'prom-uid' },
       ]);
 
@@ -272,7 +270,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when data source not found', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([
         { id: 1, name: 'Prometheus', type: 'prometheus', uid: 'prom-uid' },
       ]);
 
@@ -289,7 +287,7 @@ describe('requirements-checker.utils', () => {
 
   describe('hasPluginCHECK', () => {
     it('should check for installed plugins', async () => {
-      (ContextService.fetchPlugins as jest.Mock).mockResolvedValue([{ id: 'grafana-plugin' }]);
+      (grafanaApi.fetchPlugins as jest.Mock).mockResolvedValue([{ id: 'grafana-plugin' }]);
 
       const options: RequirementsCheckOptions = {
         requirements: 'has-plugin:grafana-plugin',
@@ -302,7 +300,7 @@ describe('requirements-checker.utils', () => {
 
   describe('pluginEnabledCHECK', () => {
     it('should check if specific plugin is enabled', async () => {
-      (ContextService.fetchPlugins as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchPlugins as jest.Mock).mockResolvedValue([
         { id: 'grafana-clock-panel', name: 'Clock Panel', enabled: true },
         { id: 'grafana-piechart-panel', name: 'Pie Chart Panel', enabled: false },
       ]);
@@ -317,7 +315,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when plugin exists but is not enabled', async () => {
-      (ContextService.fetchPlugins as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchPlugins as jest.Mock).mockResolvedValue([
         { id: 'grafana-clock-panel', name: 'Clock Panel', enabled: true },
         { id: 'grafana-piechart-panel', name: 'Pie Chart Panel', enabled: false },
       ]);
@@ -333,7 +331,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when plugin does not exist', async () => {
-      (ContextService.fetchPlugins as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchPlugins as jest.Mock).mockResolvedValue([
         { id: 'grafana-clock-panel', name: 'Clock Panel', enabled: true },
       ]);
 
@@ -350,7 +348,7 @@ describe('requirements-checker.utils', () => {
 
   describe('hasDashboardNamedCHECK', () => {
     it('should check for dashboard by name', async () => {
-      (ContextService.fetchDashboardsByName as jest.Mock).mockResolvedValue([{ title: 'Test Dashboard' }]);
+      (grafanaApi.fetchDashboardsByName as jest.Mock).mockResolvedValue([{ title: 'Test Dashboard' }]);
 
       const options: RequirementsCheckOptions = {
         requirements: 'has-dashboard-named:Test Dashboard',
@@ -628,7 +626,7 @@ describe('requirements-checker.utils', () => {
 
   describe('hasDatasourcesCHECK', () => {
     it('should check for any data sources', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([{ name: 'Test DS' }]);
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([{ name: 'Test DS' }]);
 
       const options: RequirementsCheckOptions = {
         requirements: 'has-datasources',
@@ -639,7 +637,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should fail when no data sources exist', async () => {
-      (ContextService.fetchDataSources as jest.Mock).mockResolvedValue([]);
+      (grafanaApi.fetchDataSources as jest.Mock).mockResolvedValue([]);
 
       const options: RequirementsCheckOptions = {
         requirements: 'has-datasources',
@@ -923,7 +921,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should support plugin checks as postconditions', async () => {
-      (ContextService.fetchPlugins as jest.Mock).mockResolvedValue([
+      (grafanaApi.fetchPlugins as jest.Mock).mockResolvedValue([
         { id: 'grafana-clock-panel', name: 'Clock Panel', enabled: true },
       ]);
 
@@ -936,7 +934,7 @@ describe('requirements-checker.utils', () => {
     });
 
     it('should support dashboard checks as postconditions', async () => {
-      (ContextService.fetchDashboardsByName as jest.Mock).mockResolvedValue([{ title: 'New Dashboard' }]);
+      (grafanaApi.fetchDashboardsByName as jest.Mock).mockResolvedValue([{ title: 'New Dashboard' }]);
 
       const options: RequirementsCheckOptions = {
         requirements: 'has-dashboard-named:New Dashboard',
@@ -1094,9 +1092,9 @@ describe('CHECK_HANDLERS routing parity', () => {
 
 describe('condition array execution', () => {
   it('passes an entire dashboard name to the Grafana check', async () => {
-    jest.mocked(ContextService.fetchDashboardsByName).mockResolvedValue([{ title: 'CPU, memory' }] as never);
+    jest.mocked(grafanaApi.fetchDashboardsByName).mockResolvedValue([{ title: 'CPU, memory' }] as never);
     const result = await checkRequirements({ requirements: ['has-dashboard-named:CPU, memory'], maxRetries: 0 });
-    expect(ContextService.fetchDashboardsByName).toHaveBeenCalledWith('CPU, memory', { throwOnError: true });
+    expect(grafanaApi.fetchDashboardsByName).toHaveBeenCalledWith('CPU, memory', { throwOnError: true });
     expect(result.error).toHaveLength(1);
     expect(result.pass).toBe(true);
   });
@@ -1104,10 +1102,10 @@ describe('condition array execution', () => {
 
 describe('structured check verdicts', () => {
   it('distinguishes missing dashboards from unavailable reads', async () => {
-    jest.mocked(ContextService.fetchDashboardsByName).mockResolvedValueOnce([]);
+    jest.mocked(grafanaApi.fetchDashboardsByName).mockResolvedValueOnce([]);
     const missing = await checkPostconditions({ requirements: ['has-dashboard-named:Example'], maxRetries: 0 });
     expect(missing).toMatchObject({ pass: false, verdict: 'unsatisfied' });
-    jest.mocked(ContextService.fetchDashboardsByName).mockRejectedValueOnce(new Error('Offline'));
+    jest.mocked(grafanaApi.fetchDashboardsByName).mockRejectedValueOnce(new Error('Offline'));
     const unavailable = await checkPostconditions({ requirements: ['has-dashboard-named:Example'], maxRetries: 0 });
     expect(unavailable).toMatchObject({ pass: false, verdict: 'unavailable' });
   });
